services/agent/requirements.txt
pipecat-ai[deepgram,elevenlabs,livekit]==0.0.67
aioredis==2.0.1
prometheus-client==0.20.0
